pyproject.toml
[tool.poetry]
name = "camera_manager"
version = "0.1.0"
description = ""
authors = ["Santiago Salgado <speedy@gonzales>"]

[tool.poetry.dependencies]
python = "^3.7"
ipdb = "^0.13.2"
django = "^3.0.5"
python-dotenv = "^0.12.0"
django_extensions = "^2.2.9"
django-crispy-forms = "^1.9.0"
django-chartjs = "^2.0.0"
psycopg2-binary = "^2.8.5"
opencv-python = "^4.2.0"
mrcnn = "^0.2"
geopy = "^1.21.0"
requests = "^2.23.0"
django_tables2 = "^2.3.1"
django-debug-toolbar = "^2.2"
djangorestframework = "^3.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
